requirements.txt
torch
torchvision
timm
numpy
pyyaml 